requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
AutoChromedriver==2.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
h11==0.14.0
idna==3.7
instaloader==4.11
multidict==6.0.5
outcome==1.3.0.post0
PySocks==1.7.1
pyTelegramBotAPI==4.17.0
python-i18n==0.3.9
pytube==15.0.0
PyYAML==6.0.1
requests==2.31.0
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tiktok-scraper==1.0.3
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
yarl==1.9.4
